Modify Ability Score Dialog
Hopefully this will enable the ability score dialog to be modularised
somewhat, and actually used for level ups (or however we decide to
allocate ability scores) later.

This also removes a graphical nuisance. Originally anytime the state in
the stats reducer would change, the stats bar would flash grey. Moving
the abilities for the dialog into the dialog reducer removes this,
instead opting to flash the stats bar when we actually modify the stats
completely for the player (once they click 'Confirm').