Load With Initial State
I've added { ...initialState } so that if the person uses an outdated
save file which was saved before a new state object was added to a
reducer, then the default value will be used.

The potential problem here is that it could have some sort of adverse
effect due to it possibly having changed several times throughout play
(for example, they would have missed out on the initial ability score
points from the beginning of the game, or some points from levelling
up). Perhaps we could change this instead to display a warning/error for
incompatible save files?