Add Selective Dice
This is added slightly differently to how it is defined on the wikipedia
page mentioned.

The lowest selector is like 'adxly', where a is the number of dice to
roll, x is the sides of the dice, and y is the number of 'low' values to
drop.

The highest selector is like 'adxhy', where a is the number of dice to
roll, x is the sides of the dies, and y is the number of 'high' values
to keep.